+++ docs/conf.py
@@ -57,6 +57,7 @@ source_suffix = '.rst'
 source_encoding = 'utf-8'
 master_doc = 'index'
 suppress_warnings = ['image.nonlocal_uri']
+linkcheck_ignore = [r'http://localhost:\d+']
 
 project = u'NengoDL'
 authors = u'Applied Brain Research'